Moving test to another file added timeout to spawn sync
